feat(e2e): add android-only variants for migration and witnessed exchange tests
Both e2e/run-store-migration.js and e2e/run-vrc-exchange-witnessed-devices.js
required macOS/Xcode purely for an iOS simulator or iPhone peer that has
nothing to do with what each test actually proves. Extract the shared flow
logic into lib/storeMigrationFlow.js and lib/witnessedExchangeFlow.js
(behavior-preserving refactor of the existing scripts), then add two new
entry points that swap the peer for Android:

- run-store-migration-android-only.js: second Android emulator peer
  (ANDROID_AVD2), fully unattended like the existing android-only vrc-exchange
  variant.
- run-vrc-exchange-witnessed-android-only-devices.js: second physical Android
  phone (ANDROID_UDID/ANDROID_UDID2) rather than an emulator, since hardware
  attestation is the whole point of this test and emulators can't do it.

Also fixes dangling docs/spikes/witnessed-e2e-spec.md references (that file
never existed) and adds the witnessed-devices test to the README command
table, which had been missing entirely.
