+++ e2e/package.json
@@ -9,8 +9,10 @@
     "vrc-exchange:android-only": "PLATFORMS=android,android node run-vrc-exchange.js",
     "vrc-exchange:devices": "node run-vrc-exchange-devices.js",
     "store-migration": "node run-store-migration.js",
+    "store-migration:android-only": "node run-store-migration-android-only.js",
     "onboarding-smoke": "node run-onboarding-smoke.js",
-    "vrc-exchange:witnessed:devices": "node run-vrc-exchange-witnessed-devices.js"
+    "vrc-exchange:witnessed:devices": "node run-vrc-exchange-witnessed-devices.js",
+    "vrc-exchange:witnessed:android-only": "node run-vrc-exchange-witnessed-android-only-devices.js"
   },
   "dependencies": {
     "webdriverio": "^9.12.0"
