+++ e2e/run-vrc-exchange-witnessed-devices.js
@@ -1,55 +1,34 @@
 /**
- * WITNESSED two-wallet VRC exchange on REAL DEVICES (attended).
+ * WITNESSED two-wallet VRC exchange on REAL DEVICES (attended): an Android
+ * phone paired with an iPhone. Flow lives in lib/witnessedExchangeFlow.js —
+ * see that file's header for the phase breakdown; this script only wires up
+ * device discovery and the two sessions.
  *
  * Same hardware-attested exchange as run-vrc-exchange-devices.js, but both
  * wallets first connect to a locally-run witness server, so the exchange
  * auto-routes through the witness and each wallet ends up with a Verifiable
  * Witness Credential (VWC) in addition to the peer VRC.
  *
- *   start witness (LAN-reachable) → fresh installs → onboarding →
- *   BOTH wallets connect to the witness → invitation → bidirectional,
- *   hardware-attested, witnessed VRC exchange → assert VRC + VWC on both.
- *
  * With the DI cryptosuite work, this proves the witness issues a
- * DataIntegrityProof/eddsa-rdfc-2022 VWC and both wallets store it — the last
- * uncovered cell of the DI matrix. See docs/spikes/witnessed-e2e-spec.md and
- * docs/CRYPTO_SUITE_FOLLOWUP.md.
+ * DataIntegrityProof/eddsa-rdfc-2022 VWC and both wallets store it. See
+ * docs/CRYPTO_SUITE_FOLLOWUP.md and e2e/README.md.
  *
  * ATTENDED: satisfy the OS biometric/PIN prompts at the OPERATOR banners.
- * Both phones + the Mac must share a Wi-Fi (phones reach the witness at the
- * Mac's LAN IP). Override the IP with WITNESS_HOST_IP if auto-detect is wrong.
+ * The witness runs behind a cloudflared HTTPS tunnel, so no shared LAN is
+ * needed between the phones and the machine running this script.
  *
  * Usage: npm run vrc-exchange:witnessed:devices
  *        (or: yarn e2e:vrc:witnessed:devices from repo root)
  *
- * NOTE: the device-discovery / preflight / metro / log-dump helpers are
- * duplicated from run-vrc-exchange-devices.js on purpose — keeps each runner
- * self-contained (repo convention) and leaves the proven direct-exchange
- * runner untouched.
+ * No macOS/Xcode? See run-vrc-exchange-witnessed-android-only-devices.js —
+ * same flow, a second physical Android phone stands in for the iPhone
+ * (preserves hardware-attestation coverage; two-emulator variants can't,
+ * since emulators can't do hardware attestation).
  */
-import { execSync, spawn } from "node:child_process";
-import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
-import net from "node:net";
+import { execSync } from "node:child_process";
+import { existsSync, readFileSync } from "node:fs";
 
-import {
-  createSession,
-  ensureAppium,
-  stopAppium,
-  screenshot,
-  dumpSource,
-  sleep,
-} from "./lib/driver.js";
-import {
-  acceptCredentialOfferFromChat,
-  acceptInvitationViaPaste,
-  assertVrcReceived,
-  assertContactShields,
-  completeOnboarding,
-  connectToWitness,
-  enableHardwareAttestation,
-  showRelationshipInvitation,
-} from "./lib/flows.js";
-import { startWitness } from "./lib/witness.js";
+import { createSession } from "./lib/driver.js";
 import {
   ANDROID_APK,
   ANDROID_UDID,
@@ -58,8 +37,7 @@ import {
   androidDeviceCaps,
   iosDeviceCaps,
 } from "./lib/config.js";
-
-const WITNESS_NAME = process.env.WITNESS_NAME || "e2e-witness";
+import { runWitnessedExchange, dumpAndroidWitnessLogs } from "./lib/witnessedExchangeFlow.js";
 
 // ---------- device discovery ----------
 
@@ -123,8 +101,6 @@ function detectIosUdid() {
   return candidates[0].udid;
 }
 
-// ---------- preflight / metro ----------
-
 function preflight() {
   if (!existsSync(ANDROID_APK)) {
     throw new Error(
@@ -138,166 +114,18 @@ function preflight() {
   }
 }
 
-function portInUse(port) {
-  return new Promise((resolve) => {
-    const sock = net.connect(port, "127.0.0.1");
-    sock.once("connect", () => (sock.destroy(), resolve(true)));
-    sock.once("error", () => resolve(false));
-  });
-}
-
-let metroProc;
-async function ensureMetro() {
-  if (await portInUse(8081)) {
-    console.log("[e2e] metro already running on :8081");
-    return;
-  }
-  console.log("[e2e] starting metro (yarn start in app/)…");
-  metroProc = spawn("yarn", ["start"], {
-    cwd: new URL("../app", import.meta.url).pathname,
-    stdio: ["ignore", "ignore", "inherit"],
-  });
-  for (let i = 0; i < 60; i++) {
-    if (await portInUse(8081)) return;
-    await sleep(1000);
-  }
-  throw new Error("metro did not start within 60s");
-}
-
-function dumpAndroidWitnessLogs(udid) {
-  try {
-    mkdirSync("artifacts", { recursive: true });
-    const raw = execSync(`adb -s ${udid} logcat -d`, { maxBuffer: 64 * 1024 * 1024 }).toString();
-    const lines = raw
-      .split("\n")
-      .filter((l) => /VRC:|Attestation|BiometricSignature|Witness|VWC|proofType|cryptosuite/i.test(l));
-    const file = `artifacts/witnessed-logcat-${Date.now()}.txt`;
-    writeFileSync(file, lines.join("\n"));
-    console.log(`[e2e] android witnessed log lines saved: ${file} (${lines.length} lines)`);
-    // Surface any DI proof lines for the VWC/VRC
-    for (const l of lines) {
-      if (/proofType=DataIntegrityProof|cryptosuite/i.test(l)) {
-        console.log(`[e2e]   ${l.replace(/^.*ReactNativeJS:\s*/, "").trim().slice(0, 140)}`);
-      }
-    }
-  } catch (e) {
-    console.warn(`[e2e] logcat capture failed (non-fatal): ${e.message}`);
-  }
-}
-
-// ---------- run ----------
-
-let android, ios, witness;
+preflight();
+const androidUdid = detectAndroidUdid();
 try {
-  preflight();
-  const androidUdid = detectAndroidUdid();
-  const iosUdid = detectIosUdid();
-  console.log(`[e2e] android device: ${androidUdid}`);
-
-  try {
-    execSync(`adb -s ${androidUdid} logcat -c`);
-  } catch {
-    /* non-fatal */
-  }
-
-  await ensureMetro();
-  await ensureAppium();
-
-  // The witness runs in direct mode behind a cloudflared HTTPS tunnel: the app
-  // blocks cleartext http, and production witnesses are HTTPS (real mediators
-  // like aaleon have SSL). The tunnel mirrors that locally without the
-  // instability of routing the witness through the app's mediator.
-  witness = await startWitness({ name: WITNESS_NAME });
-
-  console.log(
-    "\n[e2e] ATTENDED WITNESSED RUN — keep both phones unlocked and within reach.\n" +
-      "[e2e] Witness is reachable via an HTTPS tunnel; no LAN needed.\n" +
-      "[e2e] Authenticate at the OPERATOR banners.\n"
-  );
-
-  android = await createSession("android", androidDeviceCaps(androidUdid));
-  ios = await createSession("ios", iosDeviceCaps(iosUdid));
-
-  await Promise.all([
-    completeOnboarding(android, { firstName: "Alice", lastName: "Anderson" }),
-    completeOnboarding(ios, { firstName: "Bob", lastName: "Baker" }),
-  ]);
-
-  await Promise.all([
-    enableHardwareAttestation(android),
-    enableHardwareAttestation(ios),
-  ]);
-
-  // Both wallets connect to the witness FIRST — if either isn't connected when
-  // the exchange starts, the 15s session-challenge timeout fires and the
-  // exchange silently falls back to direct (no VWC). Confirm BOTH connections
-  // completed via the witness's own log (no "connected" banner exists in the
-  // app — witness participation only surfaces as a VWC after the exchange).
-  await connectToWitness(android, witness.invitationUrl);
-  await connectToWitness(ios, witness.invitationUrl);
-  await witness.waitForParticipants(2, 120000);
-  console.log("[e2e] both wallets connected to the witness");
-
-  const invitationUrl = await showRelationshipInvitation(android);
-  await acceptInvitationViaPaste(ios, invitationUrl);
-
-  await Promise.all([
-    acceptCredentialOfferFromChat(android, 600000, { expectAttestation: true }),
-    acceptCredentialOfferFromChat(ios, 600000, { expectAttestation: true }),
-  ]);
-
-  await Promise.all([
-    assertVrcReceived(android, "Bob Baker"),
-    assertVrcReceived(ios, "Alice Anderson"),
-  ]);
-
-  // The culminating check: each contact must show BOTH shields together —
-  // "Secure Exchange" (device attestation on the DI VRC) AND "Verified" +
-  // Witness Records (VWC from the witness). VWC is issued after the VRC, so
-  // this polls both.
-  await Promise.all([
-    assertContactShields(android, "Bob Baker"),
-    assertContactShields(ios, "Alice Anderson"),
-  ]);
-
-  console.log(
-    "\n[e2e] ✅ WITNESSED + ATTESTED VC 2.0 (eddsa-rdfc-2022) exchange succeeded on both phones —\n" +
-      "[e2e]    each contact shows BOTH shields: Secure Exchange (attestation) + Witnessed"
-  );
-  process.exitCode = 0;
-
-  dumpAndroidWitnessLogs(androidUdid);
-} catch (err) {
-  console.error("\n[e2e] ❌ FAILED:", err.message);
-  for (const d of [android, ios].filter(Boolean)) {
-    try {
-      await screenshot(d, "failure");
-      await dumpSource(d, "failure");
-    } catch {
-      /* session may be dead */
-    }
-  }
-  try {
-    dumpAndroidWitnessLogs(detectAndroidUdid());
-  } catch {
-    /* best-effort */
-  }
-  process.exitCode = 1;
-} finally {
-  for (const d of [android, ios].filter(Boolean)) {
-    try {
-      await d.deleteSession();
-    } catch {
-      /* already gone */
-    }
-  }
-  if (witness) {
-    try {
-      await witness.stop();
-    } catch {
-      /* best-effort */
-    }
-  }
-  if (metroProc) metroProc.kill("SIGTERM");
-  stopAppium();
+  execSync(`adb -s ${androidUdid} logcat -c`);
+} catch {
+  /* non-fatal */
 }
+
+await runWitnessedExchange({
+  detectDevices: () => ({ a: androidUdid, b: detectIosUdid() }),
+  createSessionA: (udid) => createSession("android", androidDeviceCaps(udid)),
+  createSessionB: (udid) => createSession("ios", iosDeviceCaps(udid)),
+  // only session A is Android here — an iOS udid isn't reachable via `adb logcat`
+  dumpWitnessLogs: () => dumpAndroidWitnessLogs([androidUdid]),
+});
