+++ package.json
@@ -39,8 +39,10 @@
     "e2e:vrc:android-only": "cd e2e && npm run vrc-exchange:android-only",
     "e2e:vrc:devices": "cd e2e && npm run vrc-exchange:devices",
     "e2e:migration": "cd e2e && npm run store-migration",
+    "e2e:migration:android-only": "cd e2e && npm run store-migration:android-only",
     "e2e:smoke": "cd e2e && npm run onboarding-smoke",
-    "e2e:vrc:witnessed:devices": "cd e2e && npm run vrc-exchange:witnessed:devices"
+    "e2e:vrc:witnessed:devices": "cd e2e && npm run vrc-exchange:witnessed:devices",
+    "e2e:vrc:witnessed:android-only": "cd e2e && npm run vrc-exchange:witnessed:android-only"
   },
   "devDependencies": {
     "@commitlint/cli": "~19.8.1",
